refactor: tracking of pending invoice payments
To reduce load on the lightning nodes, reduce the risk of paying an
invoice twice when a node goes offline and have a better insight into
the payment success metrics of nodes, the way we track payments of
lightning invoices is refactored.
Instead of asking the node directly if it still has a pending payment,
we save that information in our database and keep track of the payment
call to node in the background without blocking any locks or background
tasks that retry paying invoices.