feat: fallback to Webhook BOLT12 invoice fetching after Websocket timeout (#953)
Some mobile clients don't close the Websocket cleanly before exiting.
That can result in us trying to talk to them via the Websocket and them
never responding. Therefore, we have a shorter timeout on the Websocket
BOLT12 invoice requests now and fallback to the Webhook in case the
client does not respond on the Websocket in time.