fix: use second routing hint in MRH
The probing logic of LND assumes that the first node of the routing hint
is a LSP (and probes that node) if all route hints have the same first
node. We trick that heuristic by adding a second routing hint with a
different public key when we add a magic routing hint to the invoice.